break up networks in tx tests and helpers
