--- riemann/tests/tx/overwinter_helpers.py
@@ -1,21 +1,0 @@
-# flake8: noqa
-
-
-# https://github.com/zcash/zips/blob/master/zip-0143.rst
-
-RAW_TX = bytes.fromhex('030000807082c403024201cfb1cd8dbf69b8250c18ef41294ca97993db546c1fe01f7e9c8e36d6a5e29d4e30a703ac6a0098421c69378af1e40f64e125946f62c2fa7b2fecbcb64b6968912a6381ce3dc166d56a1d62f5a8d7056363635353e8c7203d026af786387ae60100080063656a63ac520023752997f4ff0400075151005353656597b0e4e4c705fc05020000000000000000000000000000000076495c222f7fba1e31defa3d5a57efc2e1e9b01a035587d5fb1a38e01d94903d3c3e0ad3360c1d3710acd20b183e31d49f25c9a138f49b1a537edcf04be34a9851a7af9db6990ed83dd64af3597c04323ea51b0052ad8084a8b9da948d320dadd64f5431e61ddf658d24ae67c22c8d1309131fc00fe7f235734276d38d47f1e191e00c7a1d48af046827591e9733a97fa6b679f3dc601d008285edcbdae69ce8fc1be4aac00ff2711ebd931de518856878f73476f21a482ec9378365c8f7393c94e2885315eb4671098b79535e790fe53e29fef2b3766697ac32b4f473f468a008e72389fc03880d780cb07fcfaabe3f1a84b27db59a4a153d882d2b2103596555ed9494c6ac893c49723833ec8926c1039586a7afcf4a0d9c731e985d99589c03b838e8aaf745533ed9e8ae3a1cd074a51a20da8aba18d1dbebbc862ded42435e02476930d069896cff30eb414f727b89e001afa2fb8dc3436d75a4a6f26572504b0b2232ecb9f0c02411e52596bc5e90457e745939ffedbd12863ce71a02af117d417adb3d15cc54dcb1fce467500c6b8fb86b12b56da9c382857deecc40a98d5f2903395ee4762dd21afdbb5d47fa9a6dd984d567db2857b927b7fae2db587105415d0242789d38f50b8dbcc129cab3d17d19f3355bcf73cecb8cb8a5da01307152f13902a270572670dc82d39026c6cb4cd4b0f7f5aa2a4f5a5341ec5dd715406f2fdd2a02733f5f641c8c21862a1bafce2609d9eecfa158cfb5cd79f88008e315dc7d8388036c1782fd2795d18a763624c25fa959cc97489ce75745824b77868c53239cfbdf73caec65604037314faaceb56218c6bd30f8374ac13386793f21a9fb80ad03bc0cda4a44946c00e1b1a1df0e5b87b5bece477a709649e950060591394812951e1fe3895b8cc3d14d2cf6556df6ed4b4ddd3d9a69f53357d7767f4f5ccbdbc596631277f8fecd08cb056b95e3025b9792fff7f244fc716269b926d62e9596fa825c6bf21aff9e68625a192440ea06828123d97884806f15fa08da52754a1095e3ff1abd5ce4fddfccfc3a6128aef784a64610a89d1a7099216d0814d3a2d452431c32d411ac1cce82ad0229407bbc48985675e3f874a4533f1d63a84dfa3e0f460fe2f57e34fbc75423c3737f5b2a0615f5722db041a3ef66fa483afd3c2e19e59444a64add6df1d963f5dd5b5010d3d025f0287c4cf19c75f33d51ddddba5d657b43ee8da645443814cc7329f3e9b4e54c236c29af3923101756d9fa4bd0f7d2ddaacb6b0f86a2658e0a07a05ac5b950051cd24c47a88d13d659ba2a46ca1830816d09cd7646f76f716abec5de07fe9b523410806ea6f288f8736c23357c85f45791e1708029d9824d90704607f387a03e49bf9836574431345a7877efaa8a08e73081ef8d62cb780ab6883a50a0d470190dfba10a857f82842d3825b3d6da0573d316eb160dc0b716c48fbd467f75b780149ae8808f4e68f50c0536acddf6f1aeab016b6bc1ec144b4e553acfd670f77e755fc88e0677e31ba459b44e307768958fe3789d41c2b1ff434cb30e15914f01bc6bc2307b488d2556d7b7380ea4ffd712f6b02fe806b94569cd4059f396bf29b99d0a40e5e1711ca944f72d436a102fca4b97693da0b086fe9d2e7162470d02e0f05d4bec9512bfb3f38327296efaa74328b118c27402c70c3a90b49ad4bbc68e37c0aa7d9b3fe17799d73b841e751713a02943905aae0803fd69442eb7681ec2a05600054e92eed555028f21b6a155268a2dd6640a69301a52a38d4d9f9f957ae35af7167118141ce4c9be0a6a492fe79f1581a155fa3a2b9dafd82e650b386ad3a08cb6b83131ac300b0846354a7eef9c410e4b62c47c5426907dfc6685c5c99b7141ac626ab4761fd3f41e728e1a28f89db89ffdeca364dd2f0f0739f0534556483199c71f189341ac9b78a269164206a0ea1ce73bfb2a942e7370b247c046f8e75ef8e3f8bd821cf577491864e20e6d08fd2e32b555c92c661f19588b72a89599710a88061253ca285b6304b37da2b5294f5cb354a894322848ccbdc7c2545b7da568afac87ffa005c312241c2d57f4b45d6419f0d2e2c5af33ae243785b325cdab95404fc7aed70525cddb41872cfcc214b13232edc78609753dbff930eb0dc156612b9cb434bc4b693392deb87c530435312edcedc6a961133338d786c4a3e103f60110a16b1337129704bf4754ff6ba9fbe65951e610620f71cda8fc877625f2c5bb04cbe1228b1e886f4050afd8fe94e97d2e9e85c6bb748c0042d3249abb1342bb0eebf62058bf3de080d94611a3750915b5dc6c0b3899d41222bace760ee9c8818ded599e34c56d7372af1eb86852f2a732104bdb750739de6c2c6e0f9eb7cb17f1942bfc9f4fd6ebb6b4cdd4da2bca26fac4578e9f543405acc7d86ff59158bd0cba3aef6f4a8472d144d99f8b8d1dedaa9077d4f01d4bb27bbe31d88fbefac3dcd4797563a26b1d61fcd9a464ab21ed550fe6fa09695ba0b2f10e00000000000000000000000000000000ea6468cc6e20a66f826e3d14c5006f0563887f5e1289be1b2004caca8d3f34d6e84bf59c1e04619a7c23a996941d889e4622a9b9b1d59d5e319094318cd405ba27b7e2c084762d31453ec4549a4d97729d033460fcf89d6494f2ffd789e98082ea5ce9534b3acd60fe49e37e4f666931677319ed89f85588741b3128901a93bd78e4be0225a9e2692c77c969ed0176bdf9555948cbd5a332d045de6ba6bf4490adfe7444cd467a09075417fcc0062e49f008c51ad4227439c1b4476ccd8e97862dab7be1e8d399c05ef27c6e22ee273e15786e394c8f1be31682a30147963ac8da8d41d804258426a3f70289b8ad19d8de13be4eebe3bd4c8a6f55d6e0c373d456851879f5fbc282db9e134806bff71e11bc33ab75dd6ca067fb73a043b646a70339cab4928386786d2f24141ee120fdc34d6764eafc66880ee0204f53cc1167ed02b43a52dea3ca7cff8ef35cd8e6d7c111a68ef44bcd0c1513ad47ca61c659cc5d0a5b440f6b9f59aff66879bb6688fd2859362b182f207b3175961f6411a493bffd048e7d0d87d82fe6f990a2b0a25f5aa0111a6e68f37bf6f3ac2d26b84686e569038d99c1383597fad81193c4c1b16e6a90e2d507cdfe6fbdaa86163e9cf5de310003ca7e8da047b090db9f37952fbfee76af61668190bd52ed490e677b515d0143840307219c7c0ee7fc7bfc79f325644e4df4c0d7db08e9f0bd024943c705abff899403a605cfbc7ed746a7d3f7c37d9e8bdc433b7d79e08a12f738a8f0dbddfef2f26502f3e47d1b0fd11e6a13311fb799c79c641d9da43b33e7ad012e28255398789262275f1175be8462c01491c4d842406d0ec4282c9526174a09878fe8fdde33a29604e5e5e7b2a025d6650b97dbb52befb59b1d30a57433b0a351474444099daa371046613260cf3354cfcdada663ece824ffd7e44393886a86165ddddf2b4c41773554c86995269408b11e6737a4c447586f69173446d8e48bf84cbc000a807899973eb93c5e819aad669413f8387933ad1584aa35e43f4ecd1e2d0407c0b1b89920ffdfdb9bea51ac95b557af71b89f903f5d9848f14fcbeb1837570f544d6359eb23faf38a0822da36ce426c4a2fbeffeb0a8a2e297a9d19ba15024590e3329d9fa9261f9938a4032dd34606c9cf9f3dd33e576f05cd1dd6811c6298757d77d9e810abdb226afcaa4346a6560f8932b3181fd355d5d391976183f8d99388839632d6354f666d09d3e5629ea19737388613d38a34fd0f6e50ee5a0cc9677177f50028c141378187bd2819403fc534f80076e9380cb4964d3b6b45819d3b8e9caf54f051852d671bf8c1ffde2d1510756418cb4810936aa57e6965d6fb656a760b7f19adf96c173488552193b147ee58858033dac7cd0eb204c06490bbdedf5f7571acb2ebe76acef3f2a01ee987486dfe6c3f0a5e234c127258f97a28fb5d164a8176be946b8097d0e317287f33bf9c16f9a545409ce29b1f4273725fc0df02a04ebae178b3414fb0a82d50deb09fcf4e6ee9d180ff4f56ff3bc1d3601fc2dc90d814c3256f4967d3a8d64c83fea339c51f5a8e5801fbb97835581b602465dee04b5922c2761b54245bec0c9eef2db97d22b2b3556cc969fbb13d06509765a52b3fac54b93f421bf08e18d52ddd52cc1c8ca8adfaccab7e5cc2f4573fbbf8239bb0b8aedbf8dad16282da5c9125dba1c059d0df8abf621078f02d6c4bc86d40845ac1d59710c45f07d585eb48b32fc0167ba256e73ca3b9311c62d109497957d8dbe10aa3e866b40c0baa2bc492c19ad1e6372d9622bf163fbffeaeee796a3cd9b6fbbfa4d792f34d7fd6e763cd5859dd26833d21d9bc5452bd19515dff9f4995b35bc0c1f876e6ad11f2452dc9ae85aec01fc56f8cbfda75a7727b75ebbd6bbffb43b63a3b1b671e40feb0db002974a3c3b1a788567231bf6399ff89236981149d423802d2341a3bedb9ddcbac1fe7b6435e1479c72e7089d029e7fbbaf3cf37e9b9a6b776791e4c5e6fda57e8d5f14c8c35a2d270846b9dbe005cda16af4408f3ab06a916eeeb9c9594b70424a4c1d171295b6763b22f47f80b53ccbb904bd68fd65fbd3fbdea1035e98c21a7dbc91a9b5bc7690f05ec317c97f8764eb48e911d428ec8d861b708e8298acb62155145155ae95f0a1d1501034753146e22d05f586d7f6b4fe12dad9a17f5db70b1db96b8d9a83edadc966c8a5466b61fc998c31f1070d9a5c9a6d268d304fe6b8fd3b4010348611abdcbd49fe4f85b623c7828c71382e1034ea67bc8ae97404b0c50b2a04f559e49950afcb0ef462a2ae024b0f0224dfd73684b88c7fbe92d02b68f759c4752663cd7b97a14943649305521326bde085630864629291bae25ff8822a14c4b666a9259ad0dc42a8290ac7bc7f53a16f379f758e5de750f04fd7cad47701c8597f97888bea6fa0bf2999956fbfd0ee68ec36e4688809ae231eb8bc4369f5fe1573f57e099d9c09901bf39caac48dc11956a8ae905ead86954547c448ae43d315e669c4242da565938f417bf43ce7b2b30b1cd4018388e1a910f0fc41fb0877a5925e466819d375b0a912d4fe843b76ef6f223f0f7c894f38f7ab780dfd75f669c8c06cffa43eb47565a50e3b1fa45ad61ce9a1c4727b7aaa53562f523e73952')
-
-SCRIPT_CODE = bytes.fromhex('0153')
-PREVOUT_VALUE = bytes.fromhex('418404963b4c0100')
-
-# SIGHASH_SINGLE, index=1
-HASH_PREVOUTS = bytes.fromhex('92b8af1f7e12cb8de105af154470a2ae0a11e64a24a514a562ff943ca0f35d7f')
-HASH_OUTPUTS = bytes.fromhex('e03b74bc5187184406285bb9f03b4be510f5700c5859738434cf7b8f5bdb6772')
-HASH_JOINSPLITS = bytes.fromhex('f59e41b40f3a60be90bee2be11b0956dfff06a6d8e22668c4f215bd87b20d514')
-SIGHASH = bytes.fromhex('4d266d7abfa1ace97df1e6aae5ae63627395f08b6d7b2eec15201a35faebe1d4')
-
-# https://zcha.in/transactions/c4a4901af6ac7b7672de4e2e613206dac282af0220778201d9270f1d81e1776a
-RAW_NO_JS = bytes.fromhex('030000807082c403010000000000000000000000000000000000000000000000000000000000000000ffffffff1c03216506174d696e656420627920416e74506f6f6c3707205bd62f4dffffffff0280b2e60e0000000017a914e0a5ea1340cc6b1d6a82c06c0a9c60b9898b6ae98770f99a3b000000001976a914abe7f622bee66b2761c446167723cbc74f02756588ac000000000000000000')
-
-# http://zcashnetwork.info/tx/79ceba2b3f64ac6493d9025a6c9a090f6af4b0f405d65e50e8f16ce4d009fa44
-RAW_TWO = bytes.fromhex('030000807082c403000100286bee000000001976a914bff02a61c6184c158e99b140c4ea59487095f1d688ac0000000030650600010000000000000000104f6bee0000000098d49614ceabb49488af7cdda6e3b37bdd91c7459dc3bd9346200de61e024854ee04ab295c89c27df5976e2beddcabc66eff185abc38934756899599715fccdba11b81a178e2c181f717d5b617be3c4fe05bfed7c99125379d6578d82073103f7417d0fe88b75d9d166bdc469f9bded8cabf08dcf89509387dd552fe46110de66d27671d2d23d7eae54620ce62335283156ecaf6e7fba37351bb8817339a02cea2d827fcffe3ad42f580cb046ca46bf482ad5eff064e2a7ee11be5d23693b04e8815961e91d20f95deb47cfc4f61a05a1562d3025c6a7ca27a784004a75399039a12c684c7331a208bf5f750be4b9cc40457d9ce388a27aed860fb0261a33f45e5f82da5b5b05d3c7b205b74cf1bbe657d6f0b387fdb95a9a355945198d84c6202255fa6c6d4aadf1f1be839db6b57ee67f04ab8ce4a967f917d0aad9a5bb09e0303141731dc1dca01c7f5a844268d60b1bf194e3d14b5b029dacef96bea174b1a590b0471fc12e310ae26de4ed871e6970e4d506503dbec4167398d5beb6062dbdf15d34b1bebffd5e9d314b142f6f4461a9d1a82a9913890efcf47ece748b8beac370324f0d6ae4af50ca259a48bcb6f05d83425bae66bd6c54c48cd45c1f216f566d3022e80532053c7214996266fb59b72c779c2fe260af112839952d3818990d4b5860326d8ca9b42dbd2dca5ac432d10788343dda0f8853d81ea82e15954ed7d5e8e7a0318d0dcf1eca1125bba37c189235fd4b187add6839944a5bc3dc4bb25ef6151440312a2e903f96c6453655c798c946a5309e8f54b43e312edcb8e5afc562a3edaaf4d652d91e7a4ee25f559a7ff506d9ce02bb1e6c48c7c8bdf6192f37d24316ff6a5781a3dc19d6b7162a06315ce117b18320c1703ae286c47e1329046e0d4ed2253db1a4bac817f4e100168ab66fcc5187cd0f979f9ea8cc6f317ab9ca70e5f761a9b1a1f684bb677c0c2df7b378d6aa1e9141f9e30bd61add76946ef139180a09af324b648f4cd1eecbd7588cf6f42c50432031f340f053ae8b545a4a8d70044c1743974188e7293c3ee57714c70d3cb2b8e15bd550dfdf5f35d8e8748bd7cecccd152cd4812283e1f2932bd2fd852bcaef712fb058566f14581de3bd9332d841e66fbe9562730b876c7904428ab2139e1729789e90136b2c5d66eb16bbcb3d1cc3897f735d3c330aba358c219fb84b3d690bf69937c562b8533309e4e2da55e4f1d93e6a67d7814098801485e013b0d56ab8eae0d119a150be1f0e02b0ebcac4fa554aaea5ba70026f2c0e7db11e45caf472793798729e24e3f2b9783342a591032e318b61cd958c738cc7a8ec94ca965a87cb1789e5533fc01ab26a251bc1548e49b3c52b01f9f17edf27ac1f02b968a084c0a069310f0584d78d2c1fd8960bc1bcde314d83373ca8bd72d6330d1203493db8499ac0e8f596d2b23b8bc2096ba74f07660bf504244de211b9e7283dfbd2bc06a009c05693d06eca670fa681f1d51a33d709f376e99a3b2f08c95ba42426bd372c77ed14982b1ce0610cf05dee8158cceeea93bc56a25bc1986f932c6dc8be68cba2150d9de9f2895bec3dc55476926cc8fbe2fb401cb2c06f7588b09de390de22eff4c4d6167686e75c2940da69ce781087241a815f5c0f9529d6f21191d7269466ecd7d1b319c6209b57e70aed068150e87c65f1570e634f078494d4aaf3c9a5a5db652b38b903cb9bf42eb5b2cc501c40dc1bdf48d08596392f441470d47bd6c7a3c125628ed9879d91bd676189c81e48ae8963222907a74c268b304695c5471fd6fae182352649e2eb3dc4f4796f0ff5e3f29ee95c81293cf44decc276f581faf2b15b1f3dd9abcc9baceee8d7c42c79c1ddcd772253b2f60739a567f000333796e6c3353ee4db0f02a92de6ff3398a752d1b67049f4460c050c89074673a3bc9e6eb28cee97b87fe7b8ce8a67bc06feb98c7b126bc5271b1980d0c75a7aca02bd67893d6c71b7a9eb316955fa0a6e1bbaf3c19a32bde497cca1d2a5d4fe8159191050b2c5460cefb20037b2d86e94a5713b06c2adf5a64565c02e51315b646622ee2aba6a8a24efc5cd77c683f0d4b9cfa59e1a2c1b070a5503eca4d4a765557577a8fa1c7b239e364a8578288321645307198c7fb24d5cd67e3adf14d7d166ac9ba908cf06a816ddf40c887f500a05dc8a81197f7e9b4d7b2588dff9d23efaedbc5b4b9f91ec3bb31a380cb3cedbff30785ab863e4f7df4ffcc4be3d58d73de57d3d63b99a638e107f87ceae8f4b2966a4a4c3611c5cebfcae29a9524c1afb90f1c3bcfd48fc4f9e47e551cabde97d402bb36deed57c544584daaa7cc7deec98db08ea3146c975a8b9d1692d1a3bc96ec94fce9dcebe86316eefbc12695e2e773242326d4d2a52551df231d7e95240f80eaa37d73ab9e54eb3e3c4d2f19269db72a27c0d8435cc0317e8fdb140421125bde4e7cb4f85771b28410aa5cf22da175dd6cab6827903498956026a7bf4c4838f43e19cd730350fd0db178ce7c892899e59507e313e6a313d1a1ec31308cb49e9f945e049e68eb25a6f0e569636ff259cbe11ba848100643ff90144edfb97f1b51a87c3a1776c6662347c29293da43824db71a077dfa593fa8e63368e6d6d97dc8734aeb20c49ff7909f04')
